+++ superset-frontend/spec/javascripts/views/CRUD/welcome/DashboardTable_spec.tsx
@@ -24,7 +24,7 @@ import fetchMock from 'fetch-mock';
 import { supersetTheme, ThemeProvider } from '@superset-ui/style';
 
 import ListView from 'src/components/ListView';
-import DashboardTable from 'src/welcome/DashboardTable';
+import DashboardTable from 'src/views/CRUD/welcome/DashboardTable';
 
 // store needed for withToasts(DashboardTable)
 const mockStore = configureStore([thunk]);
