+++ superset/templates/superset/crud_views.html
@@ -22,5 +22,5 @@
 {% endblock %}
 
 {% block tail_js %}
-  {{ js_bundle("welcome") }}
+  {{ js_bundle("crud_views") }}
 {% endblock %}
